역할에 따른 배치  및 이름 변경
