+++ CoordinateCalculator/CoordinateTests/CoordinateTests.swift
@@ -11,85 +11,85 @@ import XCTest
 
 class CoordinateTests: XCTestCase {
     
-    var inputScanner: InputScanner!
+    var inputChecker: InputChecker!
+    var figureMaker: FigureMaker!
     
     override func setUp() {
         super.setUp()
         
-        inputScanner = InputScanner()
+        inputChecker = InputChecker()
+        figureMaker = FigureMaker()
     }
     
     override func tearDown() {
         super.tearDown()
         
-        inputScanner = nil
+        inputChecker = nil
+        figureMaker = nil
     }
 
     // Input test
     func testMinimumInputCoordinatePassTest() {
         let testText = "(1,1)"
-        XCTAssertEqual(try inputScanner.scan(text: testText, pattern: InputScanner.validPattern), testText)
+        XCTAssertEqual(try inputChecker.checkMatching(text: testText, with: InputChecker.validPattern), testText)
     }
     
     func testMaximumInputCoordinatePassTest() {
         let testText = "(24,24)"
-        XCTAssertEqual(try inputScanner.scan(text: testText, pattern: InputScanner.validPattern), testText)
+        XCTAssertEqual(try inputChecker.checkMatching(text: testText, with: InputChecker.validPattern), testText)
     }
     
     func testBiggerCoordinateXInputTest() {
         let testText = "(25,1)"
-        XCTAssertThrowsError(try inputScanner.scan(text: testText, pattern: InputScanner.validPattern))
+        XCTAssertThrowsError(try inputChecker.checkMatching(text: testText, with: InputChecker.validPattern))
     }
     
     func testBiggerCoordinateYInputTest() {
         let testText = "(1,25)"
-        XCTAssertThrowsError(try inputScanner.scan(text: testText, pattern: InputScanner.validPattern))
+        XCTAssertThrowsError(try inputChecker.checkMatching(text: testText, with: InputChecker.validPattern))
     }
     
     func testSmallerCoordinateXInputTest() {
         let testText = "(0,1)"
-        XCTAssertThrowsError(try inputScanner.scan(text: testText, pattern: InputScanner.validPattern))
+        XCTAssertThrowsError(try inputChecker.checkMatching(text: testText, with: InputChecker.validPattern))
     }
     
     func testSmallerCoordinateYInputTest() {
         let testText = "(1,0)"
-        XCTAssertThrowsError(try inputScanner.scan(text: testText, pattern: InputScanner.validPattern))
+        XCTAssertThrowsError(try inputChecker.checkMatching(text: testText, with: InputChecker.validPattern))
     }
-    
-    ///////////////////////////////////////////
     
     func testInvalidFormatInput() {
         let testText = "((10,10))"
-        XCTAssertThrowsError(try inputScanner.scan(text: testText, pattern: InputScanner.validPattern))
+        XCTAssertThrowsError(try inputChecker.checkMatching(text: testText, with: InputChecker.validPattern))
     }
     
     func testInvalidBracketFormatInput() {
         let testText = "{10,10}"
-        XCTAssertThrowsError(try inputScanner.scan(text: testText, pattern: InputScanner.validPattern))
+        XCTAssertThrowsError(try inputChecker.checkMatching(text: testText, with: InputChecker.validPattern))
     }
     
     func testNoBracketInput() {
         let testText = "10,10"
-        XCTAssertThrowsError(try inputScanner.scan(text: testText, pattern: InputScanner.validPattern))
+        XCTAssertThrowsError(try inputChecker.checkMatching(text: testText, with: InputChecker.validPattern))
     }
     
     func testEmptyInput() {
         let testText = ""
-        XCTAssertThrowsError(try inputScanner.scan(text: testText, pattern: InputScanner.validPattern))
+        XCTAssertThrowsError(try inputChecker.checkMatching(text: testText, with: InputChecker.validPattern))
     }
     
     
     // Coordinate test
     func testGetCoordinate() {
-        let coordinate = inputScanner.getCoordinateFrom(text: "(10,10)")
-
+        let coordinate = figureMaker.getCoordinateFrom(text: "(10,10)")
         XCTAssertEqual(coordinate, [10, 10])
     }
     
     func testMakePointFromCoordinate() {
         let expectedMyPoint = MyPoint(x: 7, y: 7)
         let testCoordinate = [7, 7]
-        let testMyPoint = inputScanner.makeMyPointFrom(coordinates: testCoordinate)
+        let testMyPoint = figureMaker.makeMyPointFrom(coordinates: testCoordinate)
         
         XCTAssertEqual(expectedMyPoint.x, testMyPoint.x)
         XCTAssertEqual(expectedMyPoint.y, testMyPoint.y)
@@ -98,7 +98,7 @@ class CoordinateTests: XCTestCase {
     func testMyPointEquatable() {
         let expectedMyPoint = MyPoint(x: 7, y: 7)
         let testCoordinate = [7, 7]
-        let testMyPoint = inputScanner.makeMyPointFrom(coordinates: testCoordinate)
+        let testMyPoint = figureMaker.makeMyPointFrom(coordinates: testCoordinate)
         
         XCTAssertEqual(expectedMyPoint, testMyPoint)
     }
